refactor(auth): session 命名を刷新し verifyAccessToken の戻り値を AuthUser に変更
- services/auth/session.ts → token.ts にリネーム
- session-service.test.ts → auth-token.test.ts にリネーム
- verifyAccessToken の戻り値を JwtPayload → AuthUser に変更（sub→id 変換を内部に閉じ込め）
- テスト内の sessionService → tokenService、sessionId → token に統一
- CLAUDE.md の sessionAuth → requireAuth に更新
